Update to v025

+++ src/Input.h
@@ -4,10 +4,10 @@
 #include <genesis.h>
 #include "KeyDef.h"
 
-bool is_KeyDown(u8 key);
-bool is_KeyUp(u8 key);
-u8 get_KeyPress(u8 KeyState);
-void set_KeyPress(u8 key, u8 KeyState);
+bool is_KeyDown(u16 key);
+bool is_KeyUp(u16 key);
+u16 get_KeyPress(u8 KeyState);
+void set_KeyPress(u16 key, u8 KeyState);
 
 void Input_Init();
 void InputTick();
